refactor(registrar): extract signer manager tests
